可以上传歌曲并保存到数据库中

+++ music.go
@@ -7,7 +7,7 @@ type music struct {
 	SongName   string
 	SongAuthor string
 	AllTime    int
-	SongSize   int
+	SongSize   int64
 	URL        string
 }
